backend/requirements.txt
Django==4.0.5
djangorestframework==3.13.1
django-cors-headers>=3.13
python-dotenv==0.20.0
yt-dlp==2022.6.29
requests==2.25.1
webvtt-py==0.4.6
psycopg2-binary>=2.9

# Deps for knox
knox>=0.1
boto3>=1.24
click>=8.1
django-rest-knox>=4.2
